requirements_gemini.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
google-generativeai>=0.8.0
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sse-starlette>=1.6.0
